isolate redirect module
