+++ packages/ng/applications/sandbox/src/app/app.module.ts
@@ -3,8 +3,7 @@ import { NgModule } from '@angular/core';
 
 import { AppRouter } from './app.router';
 import { AppComponent } from './app.component';
-import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
-import { AppInterceptor } from './app.interceptor';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 @NgModule({
 	declarations: [
@@ -12,11 +11,10 @@ import { AppInterceptor } from './app.interceptor';
 	],
 	imports: [
 		BrowserModule,
+		BrowserAnimationsModule,
 		AppRouter,
-		HttpClientModule,
 	],
 	providers: [
-		{ provide: HTTP_INTERCEPTORS, useClass: AppInterceptor, multi: true },
 	],
 	bootstrap: [AppComponent]
 })
