+++ packages/ng/applications/sandbox/src/app/redirect/index.ts
@@ -1,0 +1,1 @@
+export * from './redirect.module';
